+++ app/ui/dashboard/latest-invoices.tsx
@@ -1,13 +1,11 @@
 import { ArrowPathIcon } from "@heroicons/react/24/outline";
 import Image from "next/image";
-import { LatestInvoice } from "@/app/lib/definitions";
 import clsx from "clsx";
+import { fetchLatestInvoices } from "@/app/lib/data";
 import { lusitana } from "@/app/ui/fonts";
-export default async function LatestInvoices({
-  latestInvoices,
-}: {
-  latestInvoices: LatestInvoice[];
-}) {
+
+export default async function LatestInvoices() {
+  const latestInvoices = await fetchLatestInvoices();
   return (
     <div className="flex w-full flex-col md:col-span-4">
       <h2 className={`${lusitana.className} mb-4 text-xl md:text-2xl`}>
